feat: create a separate cicd parser copied from yaml
